feat: add toolbar for fragments, packages refactoring
