+++ android/src/main/java/com/a65apps/vbabikov/studyapp/di/AppComponent.kt
@@ -1,8 +1,8 @@
 package com.a65apps.vbabikov.studyapp.di
 
-import com.a65apps.vbabikov.studyapp.ui.HomeFragment
 import com.a65apps.vbabikov.studyapp.ui.MainActivity
-import com.a65apps.vbabikov.studyapp.viewmodels.HomeViewModel
+import com.a65apps.vbabikov.studyapp.ui.calculator.CalculatorViewModel
+import com.a65apps.vbabikov.studyapp.ui.home.HomeViewModel
 import dagger.Component
 import javax.inject.Singleton
 
@@ -15,7 +15,7 @@ import javax.inject.Singleton
 interface AppComponent {
     fun inject(activity: MainActivity)
 
-    fun inject(fragment: HomeFragment)
-
     fun inject(viewModel: HomeViewModel)
+
+    fun inject(viewModel: CalculatorViewModel)
 }
