+++ android/src/main/java/com/a65apps/vbabikov/studyapp/ui/home/HomeFragment.kt
@@ -1,21 +1,25 @@
-package com.a65apps.vbabikov.studyapp.ui
+package com.a65apps.vbabikov.studyapp.ui.home
 
+import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import com.a65apps.vbabikov.studyapp.StudyApplication
 import com.a65apps.vbabikov.studyapp.databinding.FragmentHomeBinding
-import com.a65apps.vbabikov.studyapp.viewmodels.HomeViewModel
+import com.a65apps.vbabikov.studyapp.navigation.BackButtonListener
 
-class HomeFragment : Fragment() {
+class HomeFragment : Fragment(), BackButtonListener {
     private val viewModel: HomeViewModel by viewModels()
 
     private lateinit var binding: FragmentHomeBinding
 
-//    @Inject
-//    private lateinit var router: Router
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        StudyApplication.INSTANCE.appComponent.inject(viewModel)
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -29,7 +33,11 @@ class HomeFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.buttonHomeCalc.setOnClickListener {
-            viewModel.navigateToCalc()
+            viewModel.navigateToCalculator()
         }
+    }
+
+    override fun onBackPressed() {
+        viewModel.navigateBack()
     }
 }
