+++ android/src/main/java/com/a65apps/vbabikov/studyapp/ui/home/HomeViewModel.kt
@@ -1,8 +1,7 @@
-package com.a65apps.vbabikov.studyapp.viewmodels
+package com.a65apps.vbabikov.studyapp.ui.home
 
 import androidx.lifecycle.ViewModel
 import com.a65apps.vbabikov.studyapp.Screens.CalculatorFragmentScreen
-import com.a65apps.vbabikov.studyapp.StudyApplication
 import com.github.terrakok.cicerone.Router
 import javax.inject.Inject
 
@@ -10,9 +9,7 @@ class HomeViewModel : ViewModel() {
     @Inject
     lateinit var router: Router
 
-    init {
-        StudyApplication.INSTANCE.appComponent.inject(this)
-    }
+    fun navigateToCalculator() = router.navigateTo(CalculatorFragmentScreen())
 
-    fun navigateToCalc() = router.navigateTo(CalculatorFragmentScreen())
+    fun navigateBack() = router.exit()
 }
